requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
cffi==1.15.1
charset-normalizer==2.1.0
cmake==3.22.5
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.30
Deprecated==1.2.13
fonttools==4.34.4
freetype-py==2.3.0
frozenlist==1.3.0
hyperlink==21.0.0
idna==3.3
imageio==2.19.3
incremental==21.3.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8.4
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pickle5==0.0.11
Pillow==9.2.0
PTable==0.9.2
pycparser==2.21
pyglet==1.5.26
PyMCubes==0.1.2
pymeshfix==0.16.1
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
pyvista==0.36.1
PyWavelets==1.3.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
scooby==0.6.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
tqdm==4.64.0
trimesh==3.12.8
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
vedo==2022.2.3
vtk==9.0.3
wrapt==1.14.1
wslink==1.6.6
yarl==1.7.2
zope.interface==5.4.0
